refactor: move utils into sub packages
